further renaming of chartControlsInventory - > getControlsForVizType
