+++ superset-frontend/src/components/AlteredSliceTag.jsx
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 import { Table, Tr, Td, Thead, Th } from 'reactable-arc';
 import { isEqual, isEmpty } from 'lodash';
 import { getChartControlPanelRegistry } from '@superset-ui/chart';
-import getControlsForVizType from 'src/utils/chartControlsInventory';
+import getControlsForVizType from 'src/utils/getControlsForVizType';
 import { t } from '@superset-ui/translation';
 import TooltipWrapper from './TooltipWrapper';
 import ModalTrigger from './ModalTrigger';
