+++ superset-frontend/src/utils/getControlsForVizType.js
